move html generating js to index.js
