+++ index.html
@@ -114,6 +114,6 @@ let translateTable = new TranslateTable();
 <script src="tree.js"></script>
 -->
 
-<script type="module" src="src/parse.js"></script>
+<script type="module" src="src/index.js"></script>
 </body>
 </html>
