+++ src/mysplitAtDelimiters.js
@@ -2,8 +2,8 @@
 
 import { aliases, display_math_tags, possibleattributes, tags_containing_paragraphs, hint_like } from "./data";
 import { toUnicode, subenvironments} from "./data";
-import { delimitersFromList, paragraph_peer_delimiters  } from "./parse-exports";
-import { display_math_delimiters } from "./parse-exports";
+import { delimitersFromList, paragraph_peer_delimiters  } from "./parse";
+import { display_math_delimiters } from "./parse";
 import { sanitizeXMLattributes } from "./reassemble";
 
 const findEndOfMath = function(delimiter, text, startIndex) {
