+++ src/reassemble.js
@@ -1,7 +1,6 @@
 
 // import { outputtags, debugging_output_markup, PTXdisplayoutput } from "./parse-exports";
-import { debugging_output_markup, PTXdisplayoutput } from "./parse-exports";
-import { outputtags} from "./parse";
+import { debugging_output_markup, PTXdisplayoutput, outputtags } from "./parse";
 import { convertMathSnippet } from '../../Space_Math/src/main.js';
 
 let debugtags = "STart";
